Build system. card system, government, effects, religion, and so much more accomplished in this one!!!

+++ Assets/Scripts/Empires/Empire.cs
@@ -16,7 +16,7 @@ public class Empire : ScriptableObject
 
     //Empire Wealth
     public int empireGold;
-    public int empireTechPoints;
+    public float empireTechPoints;
     
     //Empire Lands
     public List<City> empireSettlements = new List<City>();
